refactor: remove unnecessary type casting in AdminThemeService

+++ src/main/java/roomescape/service/AdminThemeService.java
@@ -21,7 +21,7 @@ public class AdminThemeService {
     }
 
     public void delete(long id) {
-        if (reservationDao.existsByThemeId((long) id)) {
+        if (reservationDao.existsByThemeId(id)) {
             throw new IllegalArgumentException("예약에 사용 중인 테마는 삭제할 수 없습니다.");
         }
         themeDao.delete(id);
